after merge and solve security issues

+++ app/api/deleteProduct/[id]/route.ts
@@ -1,3 +1,5 @@
+import { decrypt } from "@/helpers/helpers";
+import { cookies } from "next/headers";
 import { NextResponse } from "next/server";
 
 export async function DELETE(
@@ -5,15 +7,20 @@ export async function DELETE(
   { params }: { params: { id: string } }
 ) {
   const { id } = params;
-  const token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwczovL3NhdmUtYml0ZS5naG9uaW0ubWFra2FoLnNvbHV0aW9ucy9hcGkvdjEvd2Vic2l0ZS9hdXRoL3NpZ24vaW4iLCJpYXQiOjE3NTExNTQxODgsImV4cCI6MTc1MjQ1MDE4OCwibmJmIjoxNzUxMTU0MTg4LCJqdGkiOiJBcWU3WDUyVVhoVWpQWjk3Iiwic3ViIjoiMjUiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.mFV0A2QR-lkAB3XGyyEaE7qIvwHQGBiGovu01i9-cUM";
+  // get and decrypt the session cookie
+  const session = (await cookies()).get("session");
+  const encryptedSession = session?.value;
+  const authToken = encryptedSession && (await decrypt(encryptedSession)).token;
+  // error if no token is found
+  if (!authToken) throw new Error("there is no token");
 
   try {
     const response = await fetch(
-      `https://save-bite.ghonim.makkah.solutions/api/v1/website/tracking-products/${id}`,
+      `${process.env.DATABASE_URL}/tracking-products/${id}`,
       {
         method: "DELETE",
         headers: {
-          Authorization: `Bearer ${token}`,
+          Authorization: `Bearer ${authToken}`,
           "Content-Type": "application/json",
         },
       }
